scrapers/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
pymongo==4.5.0
python-dateutil==2.8.2
pandas==2.1.4
apscheduler==3.10.4
python-dotenv==1.0.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1 